pyproject.toml
[tool.poetry]
name = "astrocat"
version = "0.1.0"
description = ""
authors = ["Ruslan Babayev <rus.babaef@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.1"
psycopg = {version = "^3.2.8", extras = ["binary"]}
django-environ = "^0.12.0"
minio = "^7.2.15"
pillow = "^11.2.1"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
pylint = "^3.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
split_on_trailing_comma = true
profile = "black"
